requirements.txt
# this is the requirements.txt file...

#
# DEVELOPMENT / PRODUCTION
#

# reading environment vars:
python-dotenv
pandas
requests

# web application:
flask
# production web server (for heroku):
gunicorn

# auth (login with google via oauth):
Authlib==1.0.0

#enable google sheets reading and authentication
gspread
oauth2client
